moved miller_index_from_sites to surface.py and added algo to reduce miller_indices
