+++ pymatgen/core/tests/test_surface.py
@@ -11,7 +11,8 @@ import numpy as np
 from pymatgen.core.structure import Structure
 from pymatgen.core.lattice import Lattice
 from pymatgen.core.surface import Slab, SlabGenerator, generate_all_slabs, \
-    get_symmetrically_distinct_miller_indices, ReconstructionGenerator
+    get_symmetrically_distinct_miller_indices, ReconstructionGenerator, \
+    miller_index_from_sites
 from pymatgen.symmetry.groups import SpaceGroup
 from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
 from pymatgen.util.testing import PymatgenTest
@@ -628,6 +629,28 @@ class MillerIndexFinderTests(PymatgenTest):
         # check if we were able to produce at least one
         # termination for each distinct Miller _index
         self.assertEqual(len(miller_list), len(all_miller_list))
+
+    def test_miller_index_from_sites(self):
+        # test on a cubic system
+        m = Lattice.cubic(1).matrix
+        s1 = np.array([0.5, -1.5, 3])
+        s2 = np.array([0.5, 3.,-1.5])
+        s3 = np.array([2.5, 1.5,-4.])
+        self.assertEqual(tuple(miller_index_from_sites(m, [s1, s2, s3])),
+                         (-2,-1,-1))
+
+        # test on a hexagonal system
+        l = Lattice.hexagonal(2.319, 3.626)
+        s1 = l.get_cartesian_coords([1/3, 2/3, 5/4])
+        s2 = l.get_cartesian_coords([2/3, 1/3, 3/4])
+        s3 = l.get_cartesian_coords([2/3, 4/3, -1/4])
+        s1 = np.array([1/3, 2/3, 5/4])
+        s2 = np.array([2/3, 1/3, 3/4])
+        s3 = np.array([2/3, 4/3, -1/4])
+        print(s1,s2,s3)
+        self.assertEqual(tuple(miller_index_from_sites(l.matrix, [s1, s2, s3])),
+                         (-2,-1,-1))
+
 
 if __name__ == "__main__":
     unittest.main()
